Add file validation for upload endpoint. Add detailed logging configuration by setting in .env file.

+++ curl_jwt_test.sh
@@ -3,28 +3,62 @@
 #set -x
 set -e
 
+# Get endpoint. Assign default "all" if no value was provided.
+endpoint=${1:-"all"} 
+
+# default values for upload
+model_filename=${2:-"sample_model.pkl"}
+version=${3:-"v0"}
+
+# Check if the endpoint is 'upload'
+if [[ "$endpoint" == "upload" ]]; then
+  # Check if both model_filename and version are provided
+  model_filename=${2:-"sample_model.pkl"}
+  version=${3:-"v0"}
+  if [[ -z "$model_filename" || -z "$version" ]]; then
+    echo "Error: Both model_filename and version are required for the 'upload' endpoint."
+    exit 1
+  fi
+fi
+
 TOKEN=$(curl -s -X POST http://localhost:5000/login \
     -H "Content-Type: application/json" \
     -d '{"username": "admin", "password": "pass1"}' | jq -r '.access_token')
 
+if [[ "$endpoint" == "upload" || "$endpoint" == "all" ]]; then
 # Test /upload_model endpoint using the generated token
 curl -X POST http://localhost:5000/upload_model \
     -H "Authorization: Bearer ${TOKEN}" \
     -H "Content-Type: multipart/form-data" \
-    -F "version=2.0" \
-    -F "model_file=@sample_model.pkl"
+    -F "version=$version" \
+    -F "model_file=@$model_filename"
+fi
 
+if [[ "$endpoint" == "list" || "$endpoint" == "all" ]]; then
 curl -X GET http://localhost:5000/list_models -H "Authorization: Bearer ${TOKEN}" 
+fi
 
+if [[ "$endpoint" == "remove" || "$endpoint" == "all" ]]; then
 curl -X DELETE http://localhost:5000/remove_model -H "Authorization: Bearer ${TOKEN}" -H "Content-Type: application/json" -d '{"model_filename": "random_forest_model.pkl", "version": "v4"}'
+fi
 
+if [[ "$endpoint" == "remove" || "$endpoint" == "all" ]]; then
 curl -X DELETE http://localhost:5000/remove_model -H "Authorization: Bearer ${TOKEN}" -H "Content-Type: application/json" -d '{"model_filename": "sample_model.pkl", "version": "2.0"}'
+fi
 
+if [[ "$endpoint" == "list" || "$endpoint" == "all" ]]; then
 curl -X GET http://localhost:5000/list_models -H "Authorization: Bearer ${TOKEN}" 
+fi
 
+if [[ "$endpoint" == "retrieve" || "$endpoint" == "all" ]]; then
 curl -X GET http://localhost:5000/retrieve_model -H "Authorization: Bearer ${TOKEN}" -H "Content-Type: application/json" -d '{"model_filename": "random_forest_model.pkl", "version": "v4"}'
+fi
 
+if [[ "$endpoint" == "retrieve" || "$endpoint" == "all" ]]; then
 rm -f /tmp/tmp.pkl
 curl -X GET http://localhost:5000/retrieve_model -H "Authorization: Bearer ${TOKEN}" -H "Content-Type: application/json" -d '{"model_filename": "random_forest_model.pkl", "version": "v0"}' -o /tmp/tmp.pkl
+fi
 
+if [[ "$endpoint" == "predict" || "$endpoint" == "all" ]]; then
 curl -X POST http://localhost:5000/predict -H "Authorization: Bearer ${TOKEN}" -H "Content-Type: application/json" -d '{"data": [5.9, 3.0, 4.2, 1.5], "model_filename": "sample_model.pkl", "version": "v2"}'
+fi
